+++ src/main/java/com/piotreee/pixengine/objects/Updatable.java
@@ -1,7 +1,7 @@
 package com.piotreee.pixengine.objects;
 
 public interface Updatable {
-    default void update(double delta) {
+    default void update(float delta) {
     }
 
 }
